# docs/swagger.yaml
basePath: /v1
definitions:
  main.CreatePostPayload:
    properties:
      content:
        maxLength: 100
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        maxLength: 100
        type: string
      user_id:
        type: integer
    type: object
  main.CreateUserTokenPayload:
    properties:
      email:
        maxLength: 255
        type: string
      password:
        minLength: 3
        type: string
    required:
    - email
    - password
    type: object
  main.RegisterUserPayload:
    properties:
      email:
        maxLength: 255
        type: string
      password:
        minLength: 3
        type: string
      username:
        maxLength: 100
        type: string
    required:
    - email
    - password
    - username
    type: object
  main.UpdatePostPayload:
    properties:
      content:
        maxLength: 1000
        type: string
      title:
        maxLength: 100
        type: string
    type: object
  store.Comment:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      post_id:
        type: integer
      user:
        $ref: '#/definitions/store.User'
      user_id:
        type: integer
    type: object
  store.Post:
    properties:
      comments:
        items:
          $ref: '#/definitions/store.Comment'
        type: array
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/store.User'
      user_id:
        type: integer
      version:
        type: integer
    type: object
  store.PostWithMetadata:
    properties:
      comments:
        items:
          $ref: '#/definitions/store.Comment'
        type: array
      comments_count:
        type: integer
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/store.User'
      user_id:
        type: integer
      version:
        type: integer
    type: object
  store.Role:
    properties:
      description:
        type: string
      id:
        type: integer
      level:
        type: integer
      name:
        type: string
    type: object
  store.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      role:
        $ref: '#/definitions/store.Role'
      role_id:
        type: integer
      username:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: V7 de la API para gophersocial
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger V7 API curso Backend
paths:
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register an user
      parameters:
      - description: User Credentials
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.RegisterUserPayload'
      produces:
      - application/json
      responses:
        "200":
          description: user with token
          schema:
            type: obejct
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Registers an User
      tags:
      - Authentication
  /auth/token:
    get:
      consumes:
      - application/json
      description: Creates an user token with email and password
      parameters:
      - description: User Credentials
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.CreateUserTokenPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: obejct
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Creates an user token
      tags:
      - Authentication
  /feed:
    get:
      consumes:
      - application/json
      description: Fetches Feed to a user by ID
      parameters:
      - description: Since
        in: query
        name: since
        type: string
      - description: Until
        in: query
        name: until
        type: string
      - description: Until
        in: query
        name: limit
        required: true
        type: string
      - description: Until
        in: query
        name: offset
        required: true
        type: string
      - description: Tags
        in: query
        name: tags
        type: string
      - description: Search
        in: query
        name: search
        type: string
      - description: Search
        in: query
        name: sort
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.PostWithMetadata'
            type: array
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Fetches Feed
      tags:
      - feed
  /health:
    get:
      consumes:
      - application/json
      description: Fetches Health
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: obejct
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Fetches Health
      tags:
      - DevOps
  /posts:
    post:
      consumes:
      - application/json
      description: Creates a post
      parameters:
      - description: post data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.CreatePostPayload'
      produces:
      - application/json
      responses:
        "201":
          description: created post
          schema:
            $ref: '#/definitions/store.Post'
        "400":
          description: invalid payload
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Creates a post
      tags:
      - post
  /posts/{postID}:
    delete:
      consumes:
      - application/json
      description: Deletes a post
      parameters:
      - description: postID
        in: path
        name: postID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: deleted
          schema:
            type: string
        "400":
          description: invalid payload
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Deletes a post
      tags:
      - post
    get:
      consumes:
      - application/json
      description: Fetches a post
      parameters:
      - description: postID
        in: path
        name: postID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: ' post data'
          schema:
            $ref: '#/definitions/store.Post'
        "400":
          description: invalid payload
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Fetches a post
      tags:
      - post
    patch:
      consumes:
      - application/json
      description: Updates a post
      parameters:
      - description: postID
        in: path
        name: postID
        required: true
        type: integer
      - description: post data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.UpdatePostPayload'
      produces:
      - application/json
      responses:
        "201":
          description: updated post
          schema:
            $ref: '#/definitions/store.Post'
        "400":
          description: invalid payload
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Updates a post
      tags:
      - post
  /users/{userID}:
    get:
      consumes:
      - application/json
      description: gets an user
      parameters:
      - description: userID
        in: path
        name: userID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: user data
          schema:
            $ref: '#/definitions/store.User'
        "400":
          description: invalid payload
          schema: {}
        "404":
          description: user not found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Gets an user
      tags:
      - user
  /users/{userID}/follow:
    put:
      consumes:
      - application/json
      description: Follows an user by id
      parameters:
      - description: userID
        in: path
        name: userID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: user followed
          schema:
            type: string
        "400":
          description: invalid payload
          schema: {}
        "404":
          description: user not found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Follows an user
      tags:
      - user
  /users/{userID}/unfollow:
    put:
      consumes:
      - application/json
      description: Unfollows an user by id
      parameters:
      - description: userID
        in: path
        name: userID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: user unfollowed
          schema:
            type: string
        "400":
          description: invalid payload
          schema: {}
        "404":
          description: user not found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Unfollows an user
      tags:
      - user
  /users/activate/{token}:
    put:
      description: Activates an user using the invitation token
      parameters:
      - description: Invitation token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: user activated
          schema:
            type: string
        "404":
          description: token not found or expired
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Activates an user account
      tags:
      - user
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"